ai-stock-compass-main/backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI
google-generativeai==0.8.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Financial Data
yfinance==0.2.38

# Environment & Configuration
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Date utilities
python-dateutil==2.8.2
